Consolidated Constants to 1 class.
