+++ src/main/java/frc/robot/Robot.java
@@ -1,7 +1,7 @@
 package frc.robot;
 
 import static edu.wpi.first.units.Units.Centimeters;
-import static frc.robot.constants.AbstractConstants.CONSTANTS;
+import static frc.robot.Constants.CONSTANTS;
 
 import org.littletonrobotics.junction.LogFileUtil;
 import org.littletonrobotics.junction.LoggedRobot;
