+++ src/main/java/frc/robot/subsystems/climber/Traverser.java
@@ -1,6 +1,6 @@
 package frc.robot.subsystems.climber;
 
-import static frc.robot.constants.AbstractConstants.CONSTANTS;
+import static frc.robot.Constants.CONSTANTS;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.StartEndCommand;
