+++ src/main/java/frc/robot/subsystems/led/Leds.java
@@ -1,6 +1,6 @@
 package frc.robot.subsystems.led;
 
-import static frc.robot.constants.AbstractConstants.CONSTANTS;
+import static frc.robot.Constants.CONSTANTS;
 
 import edu.wpi.first.wpilibj.AddressableLED;
 import edu.wpi.first.wpilibj.AddressableLEDBuffer;
