+++ src/main/java/frc/robot/subsystems/shooter/Aimer.java
@@ -3,7 +3,7 @@
 package frc.robot.subsystems.shooter;
 
 import static edu.wpi.first.units.Units.Amps;
-import static frc.robot.constants.AbstractConstants.CONSTANTS;
+import static frc.robot.Constants.CONSTANTS;
 
 import java.util.function.Supplier;
